+++ R/Sim_Deter.R
@@ -70,6 +70,7 @@ Buildct <- function(mesh, SurfEch= 0.03){
 #' needed. Climate variation rely on species created with mu_gr class objects.
 #' This matrix require as many rows as time steps until equil_time.
 #' If the climate does not variate, a single row can given and will be reused.
+#' @param disturbance `r lifecycle::badge("experimental")` parameter.
 #' @param correction Choice of correction of the IPM between \code{"none"}
 #' (default) and \code{"cut"}. The second option set the last column to 0 in the
 #' IPM so that no individual can grow outside of the defines classes.
@@ -102,7 +103,9 @@ Buildct <- function(mesh, SurfEch= 0.03){
 #' \describe{
 #'  \item{n}{Distribution of density by mesh along time per ha.}
 #'  \item{N}{Sum of density per ha. (colSums for n)}
-#'  \item{BA}{Basal area of the population per ha}
+#'  \item{BAsp}{Basal area of the population per ha and species}
+#'  \item{BAstand}{Basal area of the population per ha and species when
+#'  excluding size class below dth. See Harvesting vignette.}
 #'  \item{h}{Distribution of harvest density by mesh along time per ha.}
 #'  \item{H}{Sum of harvested density per ha. (colSums for h)}
 #' }
@@ -255,7 +258,7 @@ sim_deter_forest.forest  <- function(Forest,
     }
     run_disturb <- !is.null(disturbance)
     if(run_disturb){
-        # TODO idiot proff disturbance
+        # TODO idiot proof disturbance
         t_disturb <- logical(equil_time)
         t_disturb[disturbance$t] <- TRUE
         if(any(disturbance$intensity <= 0 | disturbance$intensity > 1)){